updating style folders, env and font variables
