+++ package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "@prismicio/react": "^2.0.6",
     "babel-plugin-styled-components": "^2.0.3",
+    "dotenv": "^16.0.0",
     "gatsby": "^4.7.2",
     "gatsby-plugin-image": "^2.7.0",
     "gatsby-plugin-react-helmet": "^5.7.0",
