+++ src/styles/font_names.js
@@ -1,0 +1,9 @@
+const fonts = {
+    // primary
+    space_mono: "Space Mono",
+    footerSizeDesktop: 20,
+    buttonSize: 16,
+  };
+  
+  export default fonts;
+
