+++ src/styles/fonts.scss
@@ -3,6 +3,6 @@
 
 
 body { 
-    margin: 0px !important;
+    margin: 0 !important;
 
 }
